feat: 로그인 상태에 따른 좋아요 토글 기능 분기
