+++ apps/web/components/card/hooks/use-product-like.ts
@@ -1,6 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { PRODUCT_QUERIES } from 'app/(with-layout)/(home)/components/home-section-product';
 import { apiRequest } from 'app/api/apiRequest';
+import { getCookie } from 'utils/client-cookie';
 import { useState, useEffect } from 'react';
 import { useRouter } from 'next/navigation';
 
@@ -10,18 +11,7 @@ interface UseProductLikeProps {
 
 export function useProductLike({ initialIsLiked }: UseProductLikeProps) {
   const router = useRouter();
-  function getCookie(name: string): string | null {
-    if (typeof document === 'undefined') return null;
-
-    const value = `; ${document.cookie}`;
-    const parts = value.split(`; ${name}=`);
-    if (parts.length === 2) {
-      return parts.pop()?.split(';').shift() || null;
-    }
-    return null;
-  }
   const userToken = getCookie('AccessToken');
-  console.log(userToken);
   const [isLiked, setIsLiked] = useState(initialIsLiked);
   const queryClient = useQueryClient();
 
@@ -53,7 +43,7 @@ export function useProductLike({ initialIsLiked }: UseProductLikeProps) {
   const handleLikeClick = async (
     e: React.MouseEvent,
     productId: number,
-    userToken?: string | null
+    userToken?: string | null | undefined
   ) => {
     e.stopPropagation();
     if (!userToken) {
